web_app/visitors-tracking-portal/analytics_service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
python-multipart==0.0.9
pydantic==2.6.4
httpx==0.27.0
python-jose==3.3.0
passlib==1.7.4